# api/static/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "This is a documentation for Locale API built with Flask and Flask-RestX",
        "version": "1.0.0",
        "title": "Locale API",
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "contact": {
            "name": "Barakat Adisa",
            "url": "https://github.com/Adekemi02",
            "email": "adisabarakatadekemi@gmail.com"
        }
    },
...
    "servers": [
        {
        "url": "http://localhost:5000/api/v1",
        "description": "Local server"
        },
        {
        "url": "https://locale-api.herokuapp.com/api/v1",
        "description": "Production server"
        }
    ],
    "components": {
        "securitySchemes": {
            "Bearer Auth": {
                "type": "apiKey",
                "in": "header",
                "name": "Authorization",
                "description": "Add a JWT token to the header with **Bearer &lt;JWT token&gt;** to authorize"
            }
        },
...
        "schemas": {
            "User": {
                "type": "object",
                "properties": {
                    "username": {
                        "type": "string",
                        "description": "User name"
                    },
                    "email": {
                        "type": "string",
                        "description": "User email"
                    },
                    "password": {
                        "type": "string",
                        "description": "User password"
                    },
                    "confirm_password": {
                        "type": "string",
                        "description": "User password confirmation"
                    }
                },
                "required": [
                    "username",
                    "email",
                    "password",
                    "confirm_password"
                ]
            },
            "UserLogin": {
                "type": "object",
                "properties": {
                    "email": {
                        "type": "string",
                        "description": "User email"
                    },
                    "password": {
                        "type": "string",
                        "description": "User password"
                    }
                },
                "required": [
                    "email",
                    "password"
                ]
            },
...
        }
    },
    "paths": {
        "/auth/register": {
            "post": {
                "tags": [
                    "Authentication API"
                ],
                "summary": "Register a new user",
                "description": "Register a new user",
                "requestBody": {
                    "content": {
                        "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/User"
                        }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "User created successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "409": {
                        "description": "User already exists"
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "Authentication API"
                ],
                "summary": "Login a user",
                "description": "Login a user",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserLogin"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "User logged in successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
...
        "/auth/logout": {
            "post": {
                "tags": [
                    "Authentication API"
                ],
                "summary": "Logout a user",
                "description": "Logout a user",
                "headers": {
                    "Authorization": {
                        "description": "Bearer Token",
                        "schema": {
                            "type": "string"
                        },
                        "required": true
                    }
                },
                "responses": {
                    "200": {
                        "description": "User logged out successfully"
                    },
                    "400": {
                        "description": "Bad request"
                    }
                }
            }
        },
...
    }
}